refactor functions to individual files for easier maintenance
